Rename Policy to InsurancePolicy
